[shell]

+++ tracker.sh
@@ -1,3 +1,8 @@
 #!/bin/bash
-./build/bin/tracker &
+tracker_pid=tracker.pid
+if [ -f $tracker_pid ];then
+        echo "tracker is running."
+        exit 0
+fi
+./build/bin/tracker > /dev/null 2>&1 &
 echo $! > tracker.pid
